+++ src/json_to_dns.h
@@ -10,7 +10,7 @@ extern "C" {
 // (https://developers.google.com/speed/public-dns/docs/dns-over-https#dns_response_in_json)
 // 'out' is a buffer to write the packet to. 'olen' is buffer length in bytes.
 // Returns size of packet on success, -1 on failure.
-int json_to_dns(uint16_t tx_id, char *in, uint8_t *out, int olen);
+ssize_t json_to_dns(uint16_t tx_id, char *in, uint8_t *out, int olen);
 #ifdef __cplusplus
 }
 #endif
